only test on Py3

+++ test_cupy.sh
@@ -28,9 +28,11 @@ python ../push_coveralls.py
 coverage xml
 codecov
 
-# Run benchmark
-rm -rf cupy-perf
-git clone https://github.com/niboshi/cupy-perf.git cupy-perf
-pushd cupy-perf
-python run.py --show-gpu
-popd
+# Run benchmark on Python 3
+if [ "$(python -c 'import sys; print(sys.version_info.major)')" = "3" ]; then
+  rm -rf cupy-perf
+  git clone https://github.com/niboshi/cupy-perf.git cupy-perf
+  pushd cupy-perf
+  python run.py --show-gpu
+  popd
+fi
